RFC: Single "dagster start" command that spins up dagit and the daemon as subprocesses and requires no DAGSTER_HOME env var to be set
Summary:
It is not uncommon for users to hit a roadblock when trying to get the daemon running for the first time - they need to set the DAGSTER_HOME env var, make sure its the same between two different services, keep them running in two separate windows, etc.

This proposes a single "dagster start" command that:
- spins up both services for you as subprocesses, and spins them down when the process is interrupted
- uses a subset of the workspace args available to both dagit and dagster-daemon currently
- requires no env vars: uses your current working directory as a home folder if none is set

As part of this I reduces the log level on the gRPC servers since they were a bit overwhelming out of the box (a nice future improvement would be to share gRPC servers between the two processes). We could also discuss what kind of log output we should give for people using this command for the first time to explain what's going on

Feedback on the direction welcome, if it seems promising I will clean it up and add tests.